+++ CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org).
 
 - Improved handling for @ in Number Format Masks [PR #3344](https://github.com/PHPOffice/PhpSpreadsheet/pull/3344)
 - Improved support for locale settings and currency codes when matching formatted strings to numerics in the Calculation Engine [PR #3373](https://github.com/PHPOffice/PhpSpreadsheet/pull/3373) and [PR #3374](https://github.com/PHPOffice/PhpSpreadsheet/pull/3374) 
-- Improved support for locale settings and currency identification in the Advanced Value Binder [PR #3375](https://github.com/PHPOffice/PhpSpreadsheet/pull/3375)
+- Improved support for locale settings and matching in the Advanced Value Binder [PR #3375](https://github.com/PHPOffice/PhpSpreadsheet/pull/3375)
 
 ### Deprecated